+++ src/icons/assemble.png
